+++ src/components/program/progress/material/MaterialDetail.tsx
@@ -1,44 +1,13 @@
-import React, { useRef, useState, useEffect } from "react";
+import React, { useState, useEffect } from "react";
 import styled from "@emotion/styled";
-import { FaUserCircle, FaRegWindowClose, FaPlus } from "react-icons/fa";
-import { TextField, Input } from "@mui/material";
-import { toast } from "react-toastify";
-
+import { FaUserCircle, FaPlus } from "react-icons/fa";
+import MaterialPopup from "./materialPopup/MaterialSubmitPopup";
 import { useAppDispatch, useAppSelector } from "../../../../store/hooks";
-
-
 import {
   downloadMaterialAsync,
   loadMaterialAsync,
-  uploadMaterialAsync,
 } from "../../../../slice/program/programProgressMaterial";
-
-interface Props {
-  selectedFile: File | null;
-  setSelectedFile: React.Dispatch<React.SetStateAction<File | null>>;
-}
-
-const FileUpload = (props: Props) => {
-  const handleFileChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files && event.target.files[0];
-    props.setSelectedFile(file || null);
-  };
-
-  const handleFileSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    if (props.selectedFile) {
-      console.log("Selected file:", props.selectedFile);
-    }
-  };
-
-  return (
-    <div>
-      <form onSubmit={handleFileSubmit}>
-        <input type="file" onChange={handleFileChange} />
-      </form>
-    </div>
-  );
-};
+import MaterialDetailPopup from "./materialPopup/MaterialConentPopup";
 
 const HorizonLine = () => {
   return (
@@ -51,137 +20,6 @@ const HorizonLine = () => {
         margin: "8px 0 10px",
       }}
     ></div>
-  );
-};
-
-const MaterialPopup = () => {
-  const [isOpen, setIsOpen] = useState(true);
-  const [selectedFile, setSelectedFile] = useState<File | null>(null);
-  const imageInput = useRef<HTMLInputElement>(null);
-  const dispatch = useAppDispatch();
-  const togglePopup = () => {
-    setIsOpen(!isOpen);
-  };
-  return (
-    <div>
-      {isOpen && (
-        <MPopupbox>
-          <MPopupinner>
-            <MPopupFrom>
-              <MPopupStudent>
-                <FaUserCircle size="20" color="#777777"></FaUserCircle>
-                <p>박서영</p>
-              </MPopupStudent>
-              <FaRegWindowClose
-                cursor="pointer"
-                size="20"
-                color="#777777"
-                onClick={togglePopup}
-              ></FaRegWindowClose>
-            </MPopupFrom>
-            <HorizonLine></HorizonLine>
-            <Input
-              placeholder="자료제목입력"
-              color="secondary"
-              sx={{ width: "100%", height: "14%", border: "none" }}
-            ></Input>
-            <Input
-              placeholder="자료내용입력"
-              color="secondary"
-              sx={{
-                width: "100%",
-                height: "45%",
-                border: "none",
-                marginBottom: "0.5rem",
-              }}
-            ></Input>
-            <FileUpload
-              selectedFile={selectedFile}
-              setSelectedFile={setSelectedFile}
-            ></FileUpload>
-            <HorizonLine></HorizonLine>
-            <p
-              style={{
-                color: "#07858C",
-                cursor: "pointer",
-                marginTop: "1rem",
-                marginLeft: "47%",
-              }}
-              onClick={() => {
-                const test = {
-                  programId: 1,
-                  title: "프로그램 자료 제목1",
-                  detail: "프로그램 자료 상세내용1",
-                };
-                const formData = new FormData();
-                formData.append(
-                  "data",
-                  new Blob([JSON.stringify(test)], { type: "application/json" })
-                );
-                if (selectedFile !== null) {
-                  formData.append("file", selectedFile);
-                }
-                dispatch(uploadMaterialAsync(formData));
-                togglePopup();
-              }}
-            >
-              자료올리기
-            </p>
-          </MPopupinner>
-        </MPopupbox>
-      )}
-    </div>
-  );
-};
-
-const MaterialDetailPopup = () => {
-  const [sisOpen, ssetIsOpen] = useState(true);
-  const user_gb = useAppSelector((state) => state.login.object.user_gb);
-  const subtogglePopup = () => {
-    ssetIsOpen(!sisOpen);
-  };
-  return (
-    <div>
-      {sisOpen && (
-        <MdetailPopupbox>
-          <MdetailPopupinner>
-            <MdetailPopupFrom>
-              <MdetailPopupStudent>
-                <FaUserCircle size="20" color="#777777"></FaUserCircle>
-                <p>박서영</p>
-                <p style={{ fontSize: "0.6rem" }}>2023.03.15</p>
-              </MdetailPopupStudent>
-              <FaRegWindowClose
-                cursor="pointer"
-                size="20"
-                color="#777777"
-                onClick={subtogglePopup}
-              ></FaRegWindowClose>
-            </MdetailPopupFrom>
-            <HorizonLine></HorizonLine>
-            <p style={{ marginTop: "1rem" }}>자료제목</p>
-            <p
-              style={{
-                marginTop: "1rem",
-                marginLeft: "1rem",
-                marginBottom: "6rem",
-              }}
-            >
-              자료내용
-            </p>
-
-            {user_gb === "MENTEE" && (
-              <div>
-                <HorizonLine></HorizonLine>
-                <a href="#" style={{ color: "#FF8E41" }}>
-                  자료다운받기
-                </a>
-              </div>
-            )}
-          </MdetailPopupinner>
-        </MdetailPopupbox>
-      )}
-    </div>
   );
 };
 
@@ -341,70 +179,4 @@ const Materialtextinfo = styled.div`
   width: 40%;
 `;
 
-const MPopupbox = styled.div`
-  position: fixed;
-  top: 10%;
-  left: 0;
-  width: 100%;
-  height: 90%;
-  background-color: rgba(0, 0, 0, 0.4);
-  backdrop-filter: blur(3px);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-const MPopupFrom = styled.div`
-  display: flex;
-  flex-direction: row;
-  margin-top: 0.5rem;
-  font-size: 0.8rem;
-`;
-const MPopupinner = styled.div`
-  background-color: white;
-  width: 50%;
-  height: 35%;
-  padding: 1rem;
-  border-radius: 20px;
-`;
-const MPopupStudent = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  width: 7%;
-  margin-right: 90%;
-  color: #777777;
-`;
-const MdetailPopupbox = styled.div`
-  position: fixed;
-  top: 10%;
-  left: 0;
-  width: 100%;
-  height: 90%;
-  background-color: rgba(0, 0, 0, 0.4);
-  backdrop-filter: blur(3px);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-const MdetailPopupFrom = styled.div`
-  display: flex;
-  flex-direction: row;
-  margin-top: 0.5rem;
-  font-size: 0.8rem;
-`;
-const MdetailPopupinner = styled.div`
-  background-color: white;
-  width: 50%;
-  height: 40%;
-  padding: 1rem;
-  border-radius: 20px;
-`;
-const MdetailPopupStudent = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  width: 13%;
-  margin-right: 85%;
-  color: #777777;
-`;
 export default MaterialDetail;
