+++ src/components/program/progress/notice/NoticePopup/NoticeContentPopup.tsx
@@ -1,0 +1,93 @@
+import React, { useState } from "react";
+import styled from "@emotion/styled";
+import { FaRegWindowClose, FaUserCircle } from "react-icons/fa";
+
+const HorizonLine = () => {
+  return (
+    <div
+      style={{
+        width: "100%",
+        textAlign: "center",
+        borderBottom: "1px solid #d6d6d6",
+        lineHeight: "0.1em",
+        margin: "8px 0 10px",
+      }}
+    ></div>
+  );
+};
+
+const NoticePopup = () => {
+  const [isOpen, setIsOpen] = useState(true);
+
+  const togglePopup = () => {
+    setIsOpen(!isOpen);
+  };
+  return (
+    <div>
+      {isOpen && (
+        <NPopupbox>
+          <NPopupinner>
+            <NPopupFrom>
+              <NPopupStudent>
+                <FaUserCircle size="20" color="#777777"></FaUserCircle>
+                <p>박서영</p>
+                <p style={{ fontSize: "0.6rem" }}>2023.03.15</p>
+              </NPopupStudent>
+              <FaRegWindowClose
+                cursor="pointer"
+                size="20"
+                color="#777777"
+                onClick={togglePopup}
+              ></FaRegWindowClose>
+            </NPopupFrom>
+            <HorizonLine></HorizonLine>
+            <p style={{ marginTop: "1rem" }}>공지123제목</p>
+            <p
+              style={{
+                marginTop: "1rem",
+                marginLeft: "1rem",
+                marginBottom: "6rem",
+              }}
+            >
+              공지내용
+            </p>
+          </NPopupinner>
+        </NPopupbox>
+      )}
+    </div>
+  );
+};
+const NPopupbox = styled.div`
+  position: fixed;
+  top: 10%;
+  left: 0;
+  width: 100%;
+  height: 90%;
+  background-color: rgba(0, 0, 0, 0.4);
+  backdrop-filter: blur(3px);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+const NPopupFrom = styled.div`
+  display: flex;
+  flex-direction: row;
+  margin-top: 0.5rem;
+  font-size: 0.8rem;
+`;
+const NPopupinner = styled.div`
+  background-color: white;
+  width: 50%;
+  height: 50%;
+  padding: 1rem;
+  border-radius: 20px;
+`;
+const NPopupStudent = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 13%;
+  margin-right: 84%;
+  color: #777777;
+`;
+export default NoticePopup;
