+++ src/components/program/progress/notice/NoticePopup/NoticeEditorForm.tsx
@@ -1,11 +1,8 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import styled from "@emotion/styled";
-import { Button } from "@mui/material";
+import { Button, Input } from "@mui/material";
 import { FaUser } from "react-icons/fa";
-import { toast } from "react-toastify";
-import { useAppDispatch } from "../../../../store/hooks";
-import { uploadNoticetAsync } from "../../../../slice/program/programProgressNoticeSlice";
-import ReactQuill from "react-quill";
+import { useAppDispatch } from "../../../../../store/hooks";
 import "react-quill/dist/quill.snow.css";
 interface Props {
   subtogglePopup(): void;
@@ -23,6 +20,29 @@ const HorizonLine = () => {
     ></div>
   );
 };
+const NoticeFileUpload = () => {
+  const [selectedFile, setSelectedFile] = useState<File | null>(null);
+
+  const handleFileChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const file = event.target.files && event.target.files[0];
+    setSelectedFile(file || null);
+  };
+
+  const handleFileSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    if (selectedFile) {
+      console.log("Selected file:", selectedFile);
+    }
+  };
+
+  return (
+    <div>
+      <form onSubmit={handleFileSubmit}>
+        <input type="file" onChange={handleFileChange} />
+      </form>
+    </div>
+  );
+};
 
 const NoticeEditorForm = ({ subtogglePopup }: Props) => {
   const [value, setValue] = useState("");
@@ -30,22 +50,41 @@ const NoticeEditorForm = ({ subtogglePopup }: Props) => {
   return (
     <MyBlock>
       <TestEdinfo>
-        <p>공지제목</p>
-        <input></input>
+        <p>공지제목을 입력하세요</p>
         <TestEdinfosub>
           <FaUser></FaUser>
           <p>박서영</p>
           <p>2023.03.15</p>
         </TestEdinfosub>
       </TestEdinfo>
+      <Input
+        color="secondary"
+        sx={{ height: "4vh", width: "100%", marginTop: "2rem" }}
+      ></Input>
       <HorizonLine></HorizonLine>
       <p style={{ marginLeft: "1.5%" }}>
-        <ReactQuill theme="snow" value={value} onChange={setValue} />
+        <p>공지내용을 입력하세요</p>
+        <Input
+          color="secondary"
+          sx={{
+            width: "100%",
+            height: "24vh",
+            marginBottom: "2rem",
+            marginTop: "1rem",
+          }}
+        ></Input>
+        <NoticeFileUpload></NoticeFileUpload>
+        {/* <ReactQuill theme="snow" value={value} onChange={setValue} /> */}
       </p>
       <Button
         variant="contained"
         color="secondary"
-        sx={{ height: "2.2rem", width: "9rem", marginLeft: "40%" }}
+        sx={{
+          height: "2.2rem",
+          width: "9rem",
+          marginLeft: "40%",
+          marginTop: "1rem",
+        }}
         onClick={() => {
           console.log(setValue);
           // dispatch(
@@ -88,6 +127,6 @@ const TestEdinfosub = styled.div`
   flex-direction: row;
   width: 10%;
   justify-content: space-between;
-  margin-left: 86%;
+  margin-left: 80%;
 `;
 export default NoticeEditorForm;
