+++ src/components/program/progress/onlineclass/OnlineclassDetail.tsx
@@ -6,9 +6,8 @@ import {
   FaCircle,
   FaCaretSquareRight,
   FaRegWindowClose,
-  FaUserCircle,
 } from "react-icons/fa";
-import { TextField, Checkbox, Input } from "@mui/material";
+import { Checkbox, Input } from "@mui/material";
 import { toast } from "react-toastify";
 import { useNavigate } from "react-router-dom";
 import { useAppSelector } from "../../../../store/hooks";
