+++ src/components/program/progress/qna/QnaDetail.tsx
@@ -1,12 +1,11 @@
 import React, { useEffect, useState } from "react";
 import styled from "@emotion/styled";
-import { FaUserCircle, FaRegWindowClose, FaPlus } from "react-icons/fa";
-import TestEditorForm from "../qna/TestEditorForm";
+import { FaUserCircle, FaPlus } from "react-icons/fa";
 import QnaIcon from "./QnaIcon";
 import { useAppDispatch, useAppSelector } from "../../../../store/hooks";
 import { loadQuestionListAsync } from "../../../../slice/program/programProgressQuestion";
-import ReactQuill from "react-quill";
-import "react-quill/dist/quill.snow.css";
+import SubmitPopup from "./Popup/QnasubmitPopup";
+import Popup from "./Popup/QnaContentPopup";
 
 const HorizonLine = () => {
   return (
@@ -19,73 +18,6 @@ const HorizonLine = () => {
         margin: "8px 0 20px",
       }}
     ></div>
-  );
-};
-const SubmitPopup = () => {
-  const [sisOpen, ssetIsOpen] = useState(true);
-  const subtogglePopup = () => {
-    ssetIsOpen(!sisOpen);
-  };
-  return (
-    <div>
-      {sisOpen && (
-        <SubmitPopupbox>
-          <SubmitPopupinner>
-            <FaRegWindowClose
-              size="15"
-              style={{ marginLeft: "99%" }}
-              cursor="pointer"
-              onClick={subtogglePopup}
-            ></FaRegWindowClose>
-            <TestEditorForm subtogglePopup={subtogglePopup}></TestEditorForm>
-          </SubmitPopupinner>
-        </SubmitPopupbox>
-      )}
-    </div>
-  );
-};
-
-const Popup = () => {
-  const [isOpen, setIsOpen] = useState(true);
-
-  const togglePopup = () => {
-    setIsOpen(!isOpen);
-  };
-  return (
-    <div>
-      {isOpen && (
-        <Popupbox>
-          <Popupinner>
-            <PopupFrom>
-              <PopupStudent>
-                <FaUserCircle size="20" color="#777777"></FaUserCircle>
-                <p>박서영</p>
-                <p style={{ fontSize: "0.6rem" }}>2023.03.15</p>
-              </PopupStudent>
-              <FaRegWindowClose
-                cursor="pointer"
-                size="20"
-                color="#777777"
-                onClick={togglePopup}
-              ></FaRegWindowClose>
-            </PopupFrom>
-            <HorizonLine></HorizonLine>
-            <p>질문제목</p>
-            <p
-              style={{
-                marginTop: "0.5rem",
-                marginLeft: "1rem",
-                marginBottom: "6rem",
-              }}
-            >
-              질문내용
-            </p>
-            <HorizonLine></HorizonLine>
-            <p>질문답변</p>
-          </Popupinner>
-        </Popupbox>
-      )}
-    </div>
   );
 };
 
@@ -255,56 +187,7 @@ const QnaTotal = styled.div`
   display: flex;
   flex-direction: row;
 `;
-const Popupbox = styled.div`
-  position: fixed;
-  top: 10%;
-  left: 0;
-  width: 100%;
-  height: 90%;
-  background-color: rgba(0, 0, 0, 0.4);
-  backdrop-filter: blur(3px);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-const PopupFrom = styled.div`
-  display: flex;
-  flex-direction: row;
-  margin-top: 0.5rem;
-  font-size: 0.8rem;
-`;
-const Popupinner = styled.div`
-  background-color: white;
-  width: 50%;
-  height: 60%;
-  padding: 1rem;
-  border-radius: 20px;
-`;
-const PopupStudent = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  width: 13%;
-  margin-right: 84%;
-  color: #777777;
-`;
-const SubmitPopupbox = styled.div`
-  position: fixed;
-  top: 8%;
-  left: 0;
-  width: 74%;
-  height: 75%;
-  margin-top: 4%;
-  margin-left: 21%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-const SubmitPopupinner = styled.div`
-  background-color: white;
-  width: 100%;
-  height: 100%;
-`;
+
 const Qnatextinfo = styled.div`
   display: flex;
   flex-direction: row;
