+++ src/components/program/progress/test/TestScoreForm.tsx
@@ -1,7 +1,7 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import styled from "@emotion/styled";
 
-import { Button, Input } from "@mui/material";
+import { Input } from "@mui/material";
 import { FaUserCircle } from "react-icons/fa";
 import { toast } from "react-toastify";
